Rename posts service file
