+++ app/routes/blog/index.tsx
@@ -3,7 +3,7 @@ import { LoaderFunction } from "@remix-run/node";
 import { Link, useLoaderData } from "@remix-run/react";
 import PageCard from "../../components/PageCard";
 import { Container, H1, H2 } from "../../components/styled";
-import { getPosts, PostData } from "../../services/posts";
+import { getPosts, PostData } from "../../services/posts.server";
 
 type LoaderData = {
   posts: PostData[];
